rename tests run upgrade tests only if main or tag branch
