+++ .github/workflows/build-test-image.yaml
@@ -100,7 +100,7 @@ jobs:
       matrix:
         instances:
           - name: Free - GCP/us-central1 - Failover & Persistence
-            testFile: test_standalone_new.py
+            testFile: test_standalone.py
             tierName: free-${{ github.ref_name }}
             cloudProvider: gcp
             cloudRegion: us-central1
@@ -109,7 +109,7 @@ jobs:
             environmentId: ${{ vars.OMNISTRATE_INTERNAL_DEV_ENVIRONMENT}}
             extraParams: "--resource-key 'free' --replica-id 'node-f-0' --instance-name 'test-free-failover' --instance-description 'test-free-failover' --instance-type 'none' --storage-size '30' --rdb-config 'medium' --aof-config 'always'"
           - name: PRO/Standalone - GCP/us-central1 - Failover & Persistence
-            testFile: test_standalone_new.py
+            testFile: test_standalone.py
             tierName: pro-${{ github.ref_name }}
             cloudProvider: gcp
             cloudRegion: us-central1
@@ -118,7 +118,7 @@ jobs:
             environmentId: ${{ vars.OMNISTRATE_INTERNAL_DEV_ENVIRONMENT}}
             extraParams: "--resource-key 'standalone' --replica-id 'node-s-0' --instance-name 'test-standalone-failover' --instance-description 'test-standalone-failover' --instance-type 'e2-custom-4-8192' --storage-size '30' --rdb-config 'medium' --aof-config 'always'"
           - name: PRO/SingleZone - GCP/us-central1 - Failover & Persistence
-            testFile: test_replication_new.py
+            testFile: test_replication.py
             tierName: pro-${{ github.ref_name }}
             cloudProvider: gcp
             cloudRegion: us-central1
@@ -127,7 +127,7 @@ jobs:
             environmentId: ${{ vars.OMNISTRATE_INTERNAL_DEV_ENVIRONMENT}}
             extraParams: "--resource-key 'single-Zone' --instance-name 'test-sz-failover' --instance-description 'test-sz-failover' --instance-type 'e2-custom-4-8192' --storage-size '30' --rdb-config 'medium' --aof-config 'always'"
           - name: PRO/MultiZone - GCP/us-central1 - Failover & Persistence
-            testFile: test_replication_new.py
+            testFile: test_replication.py
             tierName: pro-${{ github.ref_name }}
             cloudProvider: gcp
             cloudRegion: us-central1
@@ -136,7 +136,7 @@ jobs:
             environmentId: ${{ vars.OMNISTRATE_INTERNAL_DEV_ENVIRONMENT}}
             extraParams: "--resource-key 'multi-Zone' --instance-name 'test-mz-failover' --instance-description 'test-mz-failover' --instance-type 'e2-custom-4-8192' --storage-size '30' --rdb-config 'medium' --aof-config 'always'"
           - name: PRO/Standalone - GCP/us-central1 - Failover & Persistence With TLS
-            testFile: test_standalone_new.py
+            testFile: test_standalone.py
             tierName: pro-${{ github.ref_name }}
             cloudProvider: gcp
             cloudRegion: us-central1
@@ -145,7 +145,7 @@ jobs:
             environmentId: ${{ vars.OMNISTRATE_INTERNAL_DEV_ENVIRONMENT}}
             extraParams: "--resource-key 'standalone' --replica-id 'node-s-0' --instance-name 'test-standalone-failover-tls' --instance-description 'test-standalone-failover-tls' --instance-type 'e2-custom-4-8192' --storage-size '30' --tls --rdb-config 'medium' --aof-config 'always'"
           - name: PRO/SingleZone - GCP/us-central1 - Failover & Persistence With TLS
-            testFile: test_replication_new.py
+            testFile: test_replication.py
             tierName: pro-${{ github.ref_name }}
             cloudProvider: gcp
             cloudRegion: us-central1
@@ -154,7 +154,7 @@ jobs:
             environmentId: ${{ vars.OMNISTRATE_INTERNAL_DEV_ENVIRONMENT}}
             extraParams: "--resource-key 'single-Zone' --instance-name 'test-sz-failover-tls' --instance-description 'test-sz-failover-tls' --instance-type 'e2-custom-4-8192' --storage-size '30' --tls --rdb-config 'medium' --aof-config 'always'"
           - name: PRO/MultiZone - GCP/us-central1 - Failover & Persistence With TLS
-            testFile: test_replication_new.py
+            testFile: test_replication.py
             tierName: pro-${{ github.ref_name }}
             cloudProvider: gcp
             cloudRegion: us-central1
@@ -163,7 +163,7 @@ jobs:
             environmentId: ${{ vars.OMNISTRATE_INTERNAL_DEV_ENVIRONMENT}}
             extraParams: "--resource-key 'multi-Zone' --instance-name 'test-mz-failover-tls' --instance-description 'test-mz-failover-tls' --instance-type 'e2-custom-4-8192' --storage-size '30' --tls --rdb-config 'medium' --aof-config 'always'"
           - name: PRO/Standalone - GCP/us-central1 - Update Memory
-            testFile: test_update_memory_new.py
+            testFile: test_update_memory.py
             tierName: pro-${{ github.ref_name }}
             cloudProvider: gcp
             cloudRegion: us-central1
@@ -172,7 +172,7 @@ jobs:
             environmentId: ${{ vars.OMNISTRATE_INTERNAL_DEV_ENVIRONMENT}}
             extraParams: "--resource-key 'standalone' --instance-name 'test-standalone-update-memory' --instance-description 'test-standalone-update-memory' --instance-type 'e2-custom-4-8192' --new-instance-type 'e2-custom-8-16384' --storage-size '30' --rdb-config 'medium' --aof-config 'always'"
           - name: PRO/SingleZone - GCP/us-central1 - Update Memory
-            testFile: test_update_memory_new.py
+            testFile: test_update_memory.py
             tierName: pro-${{ github.ref_name }}
             cloudProvider: gcp
             cloudRegion: us-central1
@@ -181,7 +181,8 @@ jobs:
             environmentId: ${{ vars.OMNISTRATE_INTERNAL_DEV_ENVIRONMENT}}
             extraParams: "--resource-key 'single-Zone' --instance-name 'test-replication-update-memory' --instance-description 'test-replication-update-memory' --instance-type 'e2-custom-4-8192' --new-instance-type 'e2-custom-8-16384' --storage-size '30' --rdb-config 'medium' --aof-config 'always'"
           - name: PRO/Standalone - GCP/us-central1 - Upgrade Version
-            testFile: test_upgrade_version_new.py
+            if: github.event_name == 'push' && (contains(github.ref, 'refs/tags/v') || contains(github.ref, 'main')
+            testFile: test_upgrade_version.py
             tierName: pro-${{ github.ref_name }}
             cloudProvider: gcp
             cloudRegion: us-central1
@@ -190,7 +191,8 @@ jobs:
             environmentId: ${{ vars.OMNISTRATE_INTERNAL_DEV_ENVIRONMENT}}
             extraParams: "--resource-key 'standalone' --instance-name 'test-standalone-upgrade' --instance-description 'test-standalone-upgrade' --instance-type 'e2-custom-4-8192' --storage-size '30' --rdb-config 'medium' --aof-config 'always'"
           - name: PRO/SingleZone - GCP/us-central1 - Upgrade Version
-            testFile: test_upgrade_version_new.py
+            if: github.event_name == 'push' && (contains(github.ref, 'refs/tags/v') || contains(github.ref, 'main')
+            testFile: test_upgrade_version.py
             tierName: pro-${{ github.ref_name }}
             cloudProvider: gcp
             cloudRegion: us-central1
@@ -214,6 +216,7 @@ jobs:
           pip install -r ./tests/requirements.txt
 
       - name: ${{ matrix.instances.name }}
+        if: ${{ matrix.instances.if }}
         env:
           SERVICE_ID: ${{ matrix.instances.serviceId }}
           ENVIRONMENT_ID: ${{ matrix.instances.environmentId }}
